Move GenerateBrokerName to submarinerbrokerinfo
It's only used in one place and doesn't need to be in a general
helpers package.
